requirements.txt
asgiref==3.2.10
backcall==0.2.0
caldav==0.7.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
cryptography==3.2.1
decorator==4.4.2
defusedxml==0.7.0rc1
Django==3.1.2
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
lxml==4.6.1
oauthlib==3.1.0
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.0
traitlets==5.0.5
urllib3==1.26.2
vobject==0.9.6.1
wcwidth==0.2.5
